refactor(cef): split platform modules
