+++ crates/tauri-runtime-cef/src/cef_impl/client/context_menu.rs
@@ -5,7 +5,7 @@
 use cef::*;
 
 wrap_context_menu_handler! {
-  pub(super) struct TauriCefContextMenuHandler {
+  pub struct TauriCefContextMenuHandler {
     devtools_enabled: bool,
   }
