+++ crates/tauri-runtime-cef/src/platform/windows/mod.rs
@@ -1,0 +1,9 @@
+// Copyright 2019-2024 Tauri Programme within The Commons Conservancy
+// SPDX-License-Identifier: Apache-2.0
+// SPDX-License-Identifier: MIT
+
+mod event_loop;
+mod icon;
+mod monitor;
+mod webview;
+mod window;
